usb: rmnet: Add support to communicate with modem over bam
Rmnet is an alternative to CDC-ECM and Windows RNDIS.
It uses Qualcomm Modem Interface(QMI) protocol to communicate
with modem. Control info goes through EP0 and DATA is on bulk
in/out. On 8960, USB Rmnet driver communicates with modem over
SMD(Shared Memory Driver) and BAM(bridge access manager). SMD
is used for control info and BAM is used for DATA exchange.

Change-Id: I98c5924bcc8d9e9b82558f18e6eca45167daa08e
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>