msm: rpc: Complete notifying the RPC Router when there is data to read
There exists a race condition where, the SDIO_XPRT read buffers are
full and the SDIO_XPRT forgets to notify the RPC Router regarding the
data available to read. This leads to buffers/data left unread forever
and hence to RPC Router stall. So notify the router properly whenever
there is data to read.

Change-Id: I97aee2802f181ce6f1851e2462c3cf964fc412cf
CRs-Fixed: 286126
CRs-Fixed: 286707
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>