CHROMIUM: drivers: brcmfmac: Rename the NVRAM image.
This change temporarily renames the NVRAM image so that subsequent board
specific ebuilds will be able to install their own images without disturbing
the kernel's version. After all the ebuilds have been merged in, another kernel
change will transition back to the original name and the kernel version of the
file will be deleted.

BUG=chrome-os-partner:3524
TEST=Built an image for Kaen and verified the wifi came up and an nvram file
with the new name was in /lib/firmware/brcm.

Signed-off-by: Gabe Black <gabeblack@google.com>

Change-Id: Ib135366cad31d7807aa623cbabef80fc97a60aa2
Reviewed-on: http://gerrit.chromium.org/gerrit/1790
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>