msm_serial_hs_lite: set uart clock rate as a part of port open/close
On some SoCs, the uart clock is trampled by other processor. To work
around this issue, set the right clock rate on port open and set it
to zero on port close.

If port is configured for console, then skip clock rate setting in port
open/close, as clock rate setting to zero in port close, effects console
functionality.

Change-Id: I94b21026609f3f3292067fda48d118388543db99
Signed-off-by: Chandra Devireddy <cdevired@codeaurora.org>