msm: clock-8x60: Vote against XO shutdown when ce_pclk is enabled
The ce_pclk is used to clock the crypto engine core in addition
to providing access to its registers. Since the crypto engine may
be active even when both the application processors are collapsed,
a vote against XO shutdown should be placed whenever the ce_pclk
is enabled to prevent its clock from being turned off.

Change-Id: If4a9aec48552a8b3c53023db5bdbb29198bca291
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>