serial: HSL: Set the clock rate in clk_en
Move setting the UART clock rate from the probe call to the
clk_en call, as setting the rate will destroy the clock
state set up by the bootloader and break earlyprintk. By
delaying the clk_set_rate call, we can continue to use
earlyprintk until the console is registered and the
early console is deactivated. The latest point at which we
can call clk_set_rate is clk_en, because we are not allowed
to enable a clock prior to setting the rate.

Change-Id: I612f67d2094384f0b932ef53a45b4f57bd173576
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>