msm: restart: Reset watchdog before re-enabling
Previously when we wanted to induce a restart, we simply disabled the
watchdog, set the bark and bite times, and re-enabled it. However, this
does not affect the watchdog count, and so sometimes it was already past
our bite time, and would not cause a reboot. So, reset the watchdog
count before re-enabling it.

Change-Id: Ied687e010466cefc2895a48d4e840a9a7575b3a3
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>