msm: subsystem-fatal-8x60: Wakeup the Q6SS after setting its NMI bit
To actually process the NMI, the Q6 needs to be awake, write
to a Q6 soft interrupt register to wake the q6 up.

This will allow the Q6 to process the NMI even when its
power collapsed, thus allowing it to flush caches and
allow logs to to be useful and coherent.

Change-Id: Ia423c9e2c2ef69cf1baf4fb8cc421bff8ea4a2fd
Signed-off-by: Vikram Mulukutla <markivx@codeaurora.org>