msm_sdcc: Additional handling during resume for charm devices
MDM9k acts as the SDIO device for Charm platform.
These type of SDIO devices can be in VDD minimization
state and we need to pull them out of low power mode
before proceeding with the resume operations.

This change calls the charm specific call back in the
resume path to take care of this. If this is not done,
MDM9k can get into a wierd state when the commands
are issued from host and MDM9k will no longer be
able to respond back to further commands even after
coming out of vdd min state subsequently.

Change-Id: I45a069a6be0d99fe8b293400e0e03f6ec19d43d1
Signed-off-by: Oluwafemi Adeyemi <aadeyemi@codeaurora.org>
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>