msm: kgsl: Rename and expand calc_idle
Rather then just calculating the idle time and returning it,
there is value in returning the raw statistics and let the caller
use them as they see fit.  Rename calc_idle to power_stats, and
use a structure to pass back the total_time and busy_time.  The
structure allows for future expansion as well as providing enough
information for others to perform interesting power and load related
calculations.

Change-Id: I5132060fd6c4c6867b39a15fccb2380108394225
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>