msm: clock-rpm: Make RPM clocks into real clocks
RPM clocks are simple clocks which map to an internal table of
clocks in the rpm clock driver. Instead of having a mapping from
one driver to another, push the details from the rpm clock driver
into the struct rpm_clk and get rid of struct rpm_clk_l.

The benefit of all this is it allows us to compile out RPM clocks
on targets that don't have certain clocks.

Change-Id: I28b363d65ffd2231e21e777e55d8ead257324f60
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>