CHROMIUM: video: tegra: Fix syncpoint increment indices
The syncpt_max value is stored in the window array at location 'i', not
the window's index 'index'.  Due to this typo, attempting to schedule a
flip where the targeted window index != the flip argument index may
result in the syncpoint not being incremented.

Fix by accessing the flip_win fields through a pointer.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

BUG=none
TEST=Use omxplayer to render directly to an overlay and verify that no
timeouts occur.
Change-Id: I4a2df3df4bfca335d2791b18b838c383b15b23a5
Reviewed-on: http://gerrit.chromium.org/gerrit/1911
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>