UPSTREAM: ARM: Use TTBR1 instead of reserved context ID
On ARMv7 CPUs that cache first level page table entries (like the
Cortex-A15), using a reserved ASID while changing the TTBR or flushing
the TLB is unsafe.

This is because the CPU may cache the first level entry as the result of
a speculative memory access while the reserved ASID is assigned. After
the process owning the page tables dies, the memory will be reallocated
and may be written with junk values which can be interpreted as global,
valid PTEs by the processor. This will result in the TLB being populated
with bogus global entries.

This patch avoids the use of a reserved context ID in the v7 switch_mm
and ASID rollover code by temporarily using the swapper_pg_dir pointed
at by TTBR1, which contains only global entries that are not tagged
with ASIDs.

Reviewed-by: Frank Rowand <frank.rowand@am.sony.com>
Tested-by: Marc Zyngier <Marc.Zyngier@arm.com>
Signed-off-by: Will Deacon <will.deacon@arm.com>
[catalin.marinas@arm.com: add LPAE support]
Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>
(cherry picked from commit 3c5f7e7b4a0346de670b08f595bd15e7eec91f97)

BUG=chromium-os:34007
TEST=general system testing

Change-Id: Ifa30aef907a8950650602b6438bacbd9256ffd13
Original-Change-Id: I853a937d0a045c0487e97d74783064329d42eaf0
Signed-off-by: Olof Johansson <olofj@chromium.org>
Original-Reviewed-on: https://gerrit.chromium.org/gerrit/31713
Origintal-Reviewed-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/31855
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>