msm: Separate support flags for idle/suspend low power modes.
The current implementation uses a unified support flag to enable a power
mode with no differentiation between idle and suspend mechanisms.
However, depending on the underlying hardware, some power modes could
be supported for idle but not for suspend and vice versa. This commit
allows supporting modes separately for idle and suspend. The sysfs
entries to enable suspend/idle mechanisms are exposed only if the
corresponding support flags are set.

Change-Id: I1814f18eb546922a1ef85dcf0bdd6d31120683b5
Signed-off-by: Maheshkumar Sivasubramanian <msivasub@codeaurora.org>