CHROMIUM: arm: tegra: remove generic disp clock divisor flag
Although disp1 and disp2 have 7.1 divisors, their corresponding
registers in the clk_rst block are not the interface to program the
divisors.  Setting the generic DIV_U71 flag may cause the code to
attempt to program the clock at a different divisor, which will confuse
any code attempting to use that clock since it isn't actually being
divided.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

Change-Id: Ic9f93d7b699e62b23887adce05f7b364904090ab
BUG=chrome-os-partner:4145
TEST=Boot Seaboard, verify that FB console and X11 come up fine, as well
as that HDMI hotplug works.
Reviewed-on: http://gerrit.chromium.org/gerrit/1874
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Olof Johansson <olofj@chromium.org>