msm: common L2 accessor routines for Krait
The L2CC on Krait has an indirect addressing mechanism.
This patch adds two functions to set and get values to
the L2CPSELR and L2CPDR.

Access to these registers may be required by other parts of the
kernel other than Perfevents and cpaccess. The r/m/w sequences
are protected by a raw spinlock

Change-Id: I8b9f8c8eb6a7f0d623182506b388451bef97f0e2
Signed-off-by: Ashwin Chaugule <ashwinc@codeaurora.org>