msm_charger: Add support for alarm to signal charge start
Since we no longer hold a wakelock when charging is complete, we
need a mechanism to wakeup the system from suspend mode. The
battery alarm uses a wakeup interrupt to deliver such a
notification. Upon wakeup, the system starts charging again if a
valid charger is present. The voltage that's used for the battery
alarm wakeup is configurable through a module parameter.

Change-Id: I6dcfa20dd3a33648fbb7dd3c5776d65401a53280
Signed-off-by: Michael Bohan <mbohan@codeaurora.org>