msm: barriers: Replace dsb() with mb()
Replace explicit dsb() calls with mb(). Now that the
generic ARM implementation defines mb() to mean (at least)
dsb(), it is appropriate to switch back to the generic
kernel version of the barriers. This is also needed for
correctness on certain targets (such as 7x27) where dsb()
is insufficient and other operations (such as outer cache
sync or writing to strongly-ordered memory) are required to
ensure proper I/O operations ordering.

Change-Id: I4738e6fc6bff2e995f71d7c1a52b271ebdd80933
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>