CHROMIUM: tegra: pwm: remove pwm_disable warning
Removing the warning from pwm_disable to match the functionality of
pwm_enable which doesn't print the warning. The warning wasn't necessary
as the code already properly handles trying to disable when already
disabled.

This situation arises is in the path to disable the panel. It is done in 2
steps: 1) set brightness down to 0 and 2) disable the panel explicitly.
The problem we were seeing is that the pwm-blc driver is smart enough where
if it sees brightness = 0, then it disables pwm. Then when we try to disable
the panel a second call to pwm_disable is done. Removing this warning makes
ARM and x86 act more consistently in this manor.

In this case, with disable called twice, the code will silently do nothing
for the second call, similiar to how pwm_enable handles the same situation.

TEST=verify that after successive calls to pwm_disable, no warning is printed.
BUG=chrome-os-partner:4330

Change-Id: Ia6bd3ea35b1a3305742b87c464624661892f0eed
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/2090
Reviewed-by: Olof Johansson <olofj@chromium.org>