msm: peripheral-reset-8960: Improve modem subsystem reset register use
When doing a shutdown_q6_untrusted() for the modem firmware Q6,
assert the reset for the whole modem subsystem. This ensures a
clean state for the next time the modem is booted. It's safe
because the modem software Q6 can't run without the firmware Q6.

Also, remove use of MSS_MODEM_RESET, since that register has no
effect on msm8960, and fix a typo in the MSS_RESET macro name.

Change-Id: I550921f0bf7fc625bc16a77908e8fe90758d714e
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>