msm: clock-local: Remove use of PLL status registers for non-FSM PLLs
The PLL status registers do not actually indicate if a PLL
is in a useable state or not, nor if their outputs are enabled.
They will indicate that the PLL is enabled whether or not it is
locked and only serve to show that an internal VCO is
oscillating.

The delays in the enable path are sufficent to allow for locking
time, so checking these status bits need not be replaced by
anything.

Change-Id: I87bc91a4932716c4e6912a1e2881ca135c058c1a
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>