board: msm7x27a: Register SDC3 device before SDC1
On devices with eMMC boot, eMMC card needs to be registered as mmcblk0
during bootup. If SD card is also inserted during bootup mmcblk0 will
be assigned to it which is in slot 1. This breaks assumption of
mmcblk0 assignment to eMMC card by android framework and target may
not boot as mmcblk0p* are not eMMC block partitions.

CRs-Fixed: 290633
Change-Id: I9b27bf7c5af222998f673b8071f5d8d770f53d2e
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>