cfg80211: Ignore downstream DEAUTH for authtry_bsses
If we receive a DEAUTH message from an AP while performing
an AUTH request, this does not constitute a reply to our AUTH
request.  In some cases an incoming DEAUTH may be due to the AP
clearing out old state from a previous authentication.
Instead we should ignore it and continue waiting for an AUTH
response.

Signed-off-by: Paul Stewart <pstew@chromium.org>

BUG=chromium-os:14765
TEST=New StrictAuth automated test (soon to be added to roaming suite)
MatFunc, SecMat

Change-Id: Ifcd0791dc3fb0479716a697afbaeb53ae16d4d40
Reviewed-on: http://gerrit.chromium.org/gerrit/2485
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Sam Leffler <sleffler@chromium.org>