msm: clock-8x60: Add measurement support for RPM-managed clocks
Add measurement support for RPM-managed bus and memory clocks.
These work the same way locally-managed measurable clocks do.

smi_clk and smi_a_clk are a bit of an exception, since the test
point runs at twice the set rate of the actual clock. To handle
this, the measured rate is divided by two (though if the clock
is examined with a scope using the test pad, it will still
appear to be running at the double-rate).

Also, the measurement hardware may not be able to measure the
double-rate SMI clocks for some hardware speed-grades. Erronous
software measurements are possible in this case, though the
clock can still be correctly observed on an oscilloscope.

CRs-Fixed: 273908
Change-Id: If889add54a5daf6c09af8eb7fa22cd46551f733b
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>