msm: clock-local: Rename struct clk_local to struct rcg_clk
All clock-type structs used the "<type>_clk" naming convention
except "clk_local". Rename it to "rcg_clk" to be consistent.
Similarly, Rename the "to_local()" macro to "to_rcg_clk()".
This also disambiguates the name of the struct, since other
stuct types can also represent locally controlled clocks.

This is done with a search-and-replace of the token names and
is not expected to result in any functional changes.

Change-Id: I5a4e3b20b39f5f5640046bbf73dce2198f3b0010
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>