msm: clock-local: Rename en_reg to ctl_reg
The register pointed to by en_reg was used for more than just
enabling clocks. In many cases, the register was also used for
manipulating MN counter resets, counter modes, and counter
enables. Rename the register to something more generic because
of this.  Also rename variables and masks that deal with these
registers to match.

This was done using a search-and-replace. No functional changes
are expected.

Change-Id: I5d6f51c4770279986053a4c0f640de08ecb3fb58
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>