mmc: msm_sdcc: Fix race during SDCC suspend and wakeup interrupt
Case1:
While SDCC suspend is in progress, just after turning off clocks
but before configuring SDC DAT1 line as a wakeup interrupt if an
Rx interrupt is raised by SDIO card, SDCC IRQ handler will not
abort ongoing suspend and the interrupt will be serviced only after
an explicit wakeup is triggerred. Fix this by holding wakelock in
any case during suspend progress.

Case2:
When SDCC suspend is in progress, if a function driver or SDIO
core layer returns an error and before exiting the SDCC suspend
handler if an Rx interrupt is raised then to abort the suspend
sdio_suspend_wlock is taken. But as there are errors in suspend
handler, DPM doesn't call sdcc driver resume call anytime. This
leads to a situation where sdio_suspend_wlock is taken but not
released. Fix this by releasing wakelock if there are any errors
in SDCC suspend handler.

CRs-Fixed: 292132
Change-Id: Id7ab78f28d652e43812ddd979352eed4544f0ecc
Signed-off-by: Sujith Reddy Thumma <sthumma@codeaurora.org>