msm: ipc: Add support for subsystem restart in MSM_IPC Router
When a remote processor restarts, the transport corresponding to that
node notifies the MSM_IPC Router of that event. MSM_IPC Router cleans
up the data structures corresponding to that transport and notifies
the local clients/servers of removal of all the concerned remote
servers/clients respectively. When the remote node comes up and
broadcasts the server information, local clients will be notified of
the arrival of the server so that they can continue communication with
the server.

Change-Id: Ifcaa7902a6f67baa653ace6dc9acd31aacb83993
Signed-off-by: Karthikeyan Ramasubramanian <kramasub@codeaurora.org>