CHROMIUM: config: build-in serio_raw
We switch the mode of the touchpad from input to serio_raw for the
synaptics driver, but we need this module loaded to do that so end
up just hardcoding the modprobe in inconvenient places in the boot.

Building the module in is better. (we've just fixed one bad experience
of loading a module before udevd starts)

BUG=chromium-os:16405
TEST=booted without modprobe line, touchpad still works

Change-Id: I53733da0713566e177a1be4fbd5bdad812c514d5
Signed-off-by: Scott James Remnant <keybuk@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/2460
Reviewed-by: Olof Johansson <olofj@chromium.org>