CHROMIUM: ARM: tegra: arthur: Set up the GPIO controlling camera power.
This change explicitly configures the GPIO controlling camera power on Arthur.
The default state is on, like it happened to be before. Also, the GPIO is
exported to userspace to make it easy to turn the camera on and off. We may
decide not to leave it that way, but it could be useful for debugging.

BUG=chrome-os-partner
TEST=Built and ran kernel, used the sysfs GPIO node to toggle the camera on
and off and saw its LED toggle.

Change-Id: Ic0e50b6b9b072a6e828a63eb64777efab23b75ab
Signed-off-by: Gabe Black <gabeblack@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/2508
Reviewed-by: Gabe Black <gabeblack@chromium.org>
Tested-by: Gabe Black <gabeblack@chromium.org>