USB: gadget: f_rmnet_smd: Debugfs for endpoint buffer monitoring
The debugfs support provides the snapshot of the number of packets
transferred between host and modem and vice-versa, the number of packets
pending during transfers, the smd channel status and the online flag
status and the smd read/write availability status which can be read
using the status file in usb_rmnet folder of debugfs.
Writing 1 to the status, resets the entries.

Change-Id: Ie06b4a3413e5a99cd94f0cf4cd4e9a2de2931802
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>