msm: 7x27A: Fix the gpdigout control register define
For 45nm PHY, gpdigout control register is at offset 36 in the ULPI
register map. But currently for 7x27A target, gpdigout is defined
as 31 instead of 36 in software. Fix this issue, by mapping gpdiout
control register to correct define (i.e 36).

Check the PHY version based on target, and if it 65nm PHY, define
gpdigout control register to 31, else define it to 36.

Change-Id: I4d90cb1931fcd3838331e4783a1b8b0c2dcdc35b
Signed-off-by: Chiranjeevi Velempati <cvelempa@codeaurora.org>