CHROMIUM: dm-verity: optionally support key-val args
BUG=chromium-os:15772
TEST=Adhoc
Build a normal image (I built for Kaen) and boot it. Nothing generates the new
argument format yet, so this code is not exercised, but the system still boots.
Build a recovery image (I also built this for Kaen), boot it, and do the
recovery procedure. Ensure the machine boots afterwards. This exercises the
optional error behavior argument.

Change-Id: Icfee3e85cc7fa18054cc3d9ce71281afd3233888
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/5141
Reviewed-by: Will Drewry <wad@chromium.org>