USB: f_rmnet_smd_sdio: Drop DTR on host suspend
Windows Hosts are not droping the DTR during suspend, but setting
the DTR as part of resume. Due to this RMNET is not getting connected.
Hence send DTR set to zero to the Modem, when the rmnet interface is
suspended.

Signed-off-by: Hemant Kumar <hemantk@codeaurora.org>
(cherry picked from commit 553778440314e1bb60212784ea27eecd646f873c)

Change-Id: I2d847e6bf2616c3e65138c859e549459e4e9fa75
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>