power: pm8921-charger: Notify bms of charging starts and stops
The bms driver needs to know when the charger driver starts charging
and when it stops charging. This information helps the bms driver to
come up with accurate value for chargecycles.

Since the charger driver doesn't have a start charging and stop charging
interrupt, rely on the state change interrupt of the charger to check if
it has transitioned from a non-charging state to charging state or
vice-versa.

Change-Id: Id2469659c8268e39f25fecd97662439bc1f888b5
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>