CHROMIUM: HACK: brcm80211: Allow scans after scanning for specific ssid
brcmfmac driver returned "busy" for all scan requests after a scan
request for a specific SSID. scan_request field was not being cleared
if the code path did not call wl_do_iscan(). Users can then not
reconnect to any SSID after disconnecting (unless they take the
interface down and then back up). This is a work-around to a bug in the closed
Broadcom firmware, the patch should be removed once they fix the firmware.

BUG=chromium-os:15434
TEST=Tested on Nvidia Seaboard + BCM4329 + 2.6.38

Change-Id: I39eea01577b5b76fab583e79998f39ac781f5f4c
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/3019
Reviewed-by: Grant Grundler <grundler@chromium.org>