msm: socinfo: Read CPU type from hardware registers
Use MIDR and MPIDR to identify the SoC type rather than
relying on values from SMEM. This enables us to use the
cpu_is macros before SMEM is initialized, and in early
bringup where SMEM does not yet exist. Only the macros for
detecting 8x60, 8960, 8064, and 8x30 are supported at this
time.

Change-Id: I37fab1d0128bd85da0bd0c7a23878a9326f8a1c2
Signed-off-by: Stepan Moskovchenko <stepanm@codeaurora.org>