msm: timer: account for shift in sched_clock
Previously we were telling the sched_clock layer that it could use 32
bits of the timer count we gave it. This is incorrect on targets in
which we shift the DGT count down due to inaccurate lower bits. This can
cause timestamp wrapping, as the sched_clock layer does not read from
the timer often enough to properly detect overflow. Fix this by
subtracting the shift of a timer from the number of bits we give to
sched_clock.

CRs-Fixed: 290340
Change-Id: I079eb674a3c7ccbd2164679da470bf7ebdaab1a7
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>