msm: timer: account for shift in sched_clock
Previously we were telling the sched_clock layer that it could use 32
bits of the timer count we gave it. This is incorrect on targets in
which we shift the DGT count down due to inaccurate lower bits. This can
cause timestamp wrapping, as the sched_clock layer does not read from
the timer often enough to properly detect overflow. Fix this by
subtracting the shift of a timer from the number of bits we give to
sched_clock.

CRs-Fixed: 290340
Signed-off-by: Jeff Ohlstein <johlstei@codeaurora.org>
(cherry picked from commit 61df9e89cc549a750fec9ede2d87ff7dbaa96462)

Change-Id: I4e4137aae90f02f2aeb9f764b91f137d2b1faf37
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>