msm: qdsp6v2: Address FTM speaker devices volume related bugs
When playing audio to ftm_spkr_l_rx device, very low gain
is observed. The reason is that default volume is set to 50%
which translate to -34 DB. Such behavior is not observed
on mission mode devices is because device volume type is not set
as we do not want application to touch device volume which is being
calibrated. speaker and heahpone mono profiles are not correct so
output of L and R CDAC are not routed to desired output mixers.
Bad profile ended up causing a lot of confusion during FTM testing

Change-Id: I6fa52f5f26f277c585aae08ea03d0afdb715d843
Signed-off-by: Patrick Lai <plai@codeaurora.org>