msm: msm_sdcc: Fix clock rate issues with DDR bus speed mode
If timing is set to DDR mode, we need to set MCLK to double
the rate requested by MMC core layer by set_ios. But
set_ios() may be called within MMC driver and in that case
clock rate passed to set_ios() may be already doubled
for DDR mode.

This change make sure that clock rate is doubled only if
doubled clock rate is not set for DDR mode.

Change-Id: Ia4cc20e324d1721e96e2a9f925251d79be6caa39
Signed-off-by: Subhash Jadavani <subhashj@codeaurora.org>