power: pm8921-bms: debug output cleanup
In the interpolate_single_lut print the name of the table when values
are out of bounds.

Change the important values to be printed in capitals in
calculate_state_of_charge functions.

dev_dbg does not follow the pr_fmt. Change to use pr_debug.

Change pr_err to pr_debug for places where the software takes corrective
action.

Change-Id: I1d62021076c424c5ffcb0ea710d7ea8f6257f988
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>