regulator: pmic8058-regulator: correct LVS pull down setting
The pull down enable bit of the LVS control registers was set the
same way as those in LDO and SMPS regulators.  I.e. set to 1 to
enable the pull down and set to 0 to disable it.  However, the pull
down bit meaning for LVS's is inverted.  Correct this in the
driver.

Change-Id: Ib568a60f54b67cd3cd8812b0084446603ce39b9c
Signed-off-by: David Collins <collinsd@codeaurora.org>