msm: board-msm8x60: unset sleep_sel flag for PM8058 L16 and L18 regulators
Set the sleep selectable flag to 0 for PMIC 8058 regulators LDO 16 and
LDO 18.  These are not manipulated via the private rpm_vreg_set_voltage
API, so sleep selectable is not needed.

A minor performance penalty is taken when sleep selectable is used
because two RPM requests must be sent (active set and sleep set)
instead of just one (active set).  Also, an additional spinlock is
taken.  Therefore, it is best to only set the sleep selectable flag
only when it is absolutely required.

Change-Id: I8243da746c82d1eb9f04b68b108a1061d41f0939
Signed-off-by: David Collins <collinsd@codeaurora.org>