msm: clock-voter: Fix disable() voting aggregation
If a clock is disabled, we want to aggregate the rate of all
other clocks besides the one that is being turned off, and then
drop the rate if the current rate is more than the rate would be
if the clock is turned off. If the clock's enabled field isn't
set to false before aggregating the other rates, we'll mistakenly
include the clocks that is being turned off in the aggregate
rate, thus not dropping the clock's rate from the votes.

Change-Id: I76d6f24a415848147872933c716274f5be1dd786
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>