CHROMIUM: dm: Add support for devices by PARTUUID/PARTNROFF.
This is a special hack for the chromium-os bootloader, which knows its own
partition UUID easily but not that of its targets.

BUG=chromium-os:17184
TEST=Adhoc
Booted a kaen with dm-verity, and 'dmesg | grep verity'. Look in particular for
verity finding a root device, and for 'adding target' with a PARTUUID line in
it.

Change-Id: Idc2791db1b9cc376090c188062c675c8dd173fd4
Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/3900