msm: sdio_dmux: manipulate skb->truesize per packet
The truesize field of a skb is used by the TCP/IP framework to determine
the memory usage of a socket, and to control the TCP window size.  Since
the original skb buffer from sdio likely contained multiple packets, the
truesize field is not accurate and results in the TCP window closing faster
than it should.

Manually change the truesize field in the cloned skbs to accurately reflect
the size of each packet so that the TCP/IP framework has accurate data to
base it's statistics on.

Change-Id: Iafecd98ac5421c75a76f9ae80c90ee0f546553a8
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>