msm: sdio_dmux: manipulate skb->truesize per packet
The truesize field of a skb is used by the TCP/IP framework to determine
the memory usage of a socket, and to control the TCP window size.  Since
the original skb buffer from sdio likely contained multiple packets, the
truesize field is not accurate and results in the TCP window closing faster
than it should.

Manually change the truesize field in the cloned skbs to accurately reflect
the size of each packet so that the TCP/IP framework has accurate data to
base it's statistics on.

CRs-fixed: 293892
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>
(cherry picked from commit 57c16e098ec4f8200d70f6f11413c9435dd49f7b)

Change-Id: I3b70279b3ecc0697cf3444590f06ffe54d8af3af
Signed-off-by: Ramakrishna Prasad N <crpn@codeaurora.org>