ASoc: msm8960: add control for setting SLIMBUS BE channels
when switching from one audio  device to another audio device with
different channel configuration, AFE SLIMBUS BE DAI channel
configuration has to be set to that of  new audio device channel
configuration before enabling AFE SLIMBUS BE DAI again for new audio
device. This has to be done to avoid the channel underflow/overflow
in Slimbus slave on codec.

Change-Id: I0022292804d4eca3dd2c3a9b3e0d40fa87a625b0
Signed-off-by: Kiran Kandi <kkandi@codeaurora.org>