CHROMIUM: gobi: fix use-after-free of setup buffer.
The buffer might outlive the function if semaphores are broken (i.e., if we have
a pending signal), so stick it in the refcounted context so it only gets cleaned
up when we're done with it.

Additionally, fix some double-frees in error cleanup paths, and document a bit.

BUG=chromium-os:15017
TEST=network_GobiUncleanDisconnect

Signed-off-by: Elly Jones <ellyjones@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/3686
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: ttuttle <ttuttle@chromium.org>
(cherry picked from commit a44802f141b9da40ac3da526c67aee93297dc54d)

Change-Id: I89f78eadc41fedc5bd9ffc12074d19098f61c3f2
Reviewed-on: http://gerrit.chromium.org/gerrit/4174
Reviewed-by: Jason Glasgow <jglasgow@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>