ASoC: WCD9310: Fix issue with headset inserted upon bootup
The headset insertion interrupt block must be enabled after the
overall WCD9310 headset insertion interrupt is enabled so that
insertion detection will work if a headset is inserted before
the device boots up.  This sequencing is necessary because the
insertion interrupt is edge triggered, so improper sequencing
will cause the codec interrupt block to miss the trigger of the
interrupt line.

Change-Id: I78118c23ac8f7770b46c28e73a7cc8bad5caac3a
CRs-fixed: 294874
Signed-off-by: Brad Rubin <brubin@codeaurora.org>