mfd: pm8058: remove mutex lock/unlock from pm8058_reset_pwr_off
Remove mutex_lock and mutex_unlock calls inside of the
pm8058_reset_pwr_off function.  There is no need to guarantee
mutual exclusion in this function because it is only called
immediately before shutting down.  Mutexes cannot be used in this
function because it must be possible to call it from an atomic
context.

Change-Id: Ie8ae821dc8a2e6561ea8a0f669bcb356d0c39af0
Signed-off-by: David Collins <collinsd@codeaurora.org>