msm: rpm-regulator-8960: change pin function value for no pin control case
Change the pin function value used when pin control is not desired.
This will ensure that a regulator is always turned on when
requested.  A request with pf=0 can be overridden by one from
another processor which requests pin control thanks to the
aggregation technique used on the RPM.

Therefore, change RPM_VREG_PIN_FN_NONE to have a value of 4, which
follows the semantics that Linux expects in the no pin control case
and define the pin function value 0 to be a "don't care" value.

Change-Id: I8cf54e14df10ff89e4de460c8c4a20e2866204a4
Signed-off-by: David Collins <collinsd@codeaurora.org>