CHROMIUM: chromeos_arm: use device tree
This change removes the old shared memory model, and uses the device
tree to pass the boot-time variables directly to userspace. Left to
handle is just the in-kernel use of being able to read and write the
"nvram" block on MMC to mark that the system needs recovery, etc.

Also, remove the locally-cashed copy of the nonvolatile storage from
the driver, since userspace does direct modification of the block by
opening the device and reading/writing from it. If a local copy is kept,
it ends up getting out of sync with the state set from userspace.

This requires a bit of trickery with flushing the buffer cache to avoid
the userspace tool from getting stale data (and vice versa).

BUG=chromium-os:16665
TEST=boot with vboot on and off, check for function

Change-Id: Id374cf42088ffe86c65b126412b92441d5aa1095
Signed-off-by: Olof Johansson <olofj@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/3766
Reviewed-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Doug Anderson <dianders@chromium.org>