partition_sched_domains: Do not destroy old sched domain on cpu_up
This is safe on a cpu_up only. Although a reader may still have
access to the old scheduling domain data, the data indicates that
the new CPU is not up, and therefore the only limitation is the
new CPU will not be schedulable by that reader until that reader
receives the new data.

Change-Id: I71de2f0be9cee10759862ce710f3f744ea916a76
Signed-off-by: Maya Spivak <mspivak@codeaurora.org>