power: msm_charger: fix wakelock destruction paths
The wakelock in the msm_charger is not specific to any charger. It is held
as long as any charger is charging the battery and released when not
charging or done charging.

It is not appropriate to destroy the wakelock when a charger registers or
unregisters, there could be other chargers that would need that
wakelock functional.

Destroy the wakelock only when the msm_charger is removed.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit 84ef67ed3a99953fe29822cd590d33aaec658f11)

Change-Id: Ifd414d82d618796991ce9b95e1caf8c63a26b602
Signed-off-by: Shruthi Krishna <skrish@codeaurora.org>