CHROMIUMOS: Do not resubmit int urb if read_urb submitted
Do not resubmit int urb if read_urb submitted.  To do so could lead to
race conditions in which the int urb is resubmitted while the
int_complete is processing and urb.  In the best case this causes a
message to be delayed, in the worst case, the message arrives too
late and is ignored.

BUG=chromium-os:17530
TEST=run_remote_test suite_Cellular

Signed-off-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/3976
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Reviewed-by: ttuttle <ttuttle@chromium.org>
(cherry picked from commit 07d0c74ec8c0e3673e15c207ca5aefa6ea76e79f)

Change-Id: I53d8ccc942b51cc3d0ec5d35aa5fa8fedb1382c9
Reviewed-on: http://gerrit.chromium.org/gerrit/4094
Reviewed-by: ttuttle <ttuttle@chromium.org>
Tested-by: ttuttle <ttuttle@chromium.org>
Reviewed-by: Elly Jones <ellyjones@chromium.org>
Tested-by: Jason Glasgow <jglasgow@chromium.org>