USB: msm_otg: Simplify ACA state machine to support standard ACA
A Standard ACA has a Standard-A receptacle on the Accessory Port,
and can only be attached to a B-device.  RID_A and RID_GND states
are only possible with Standard ACA.  As RID_A to RID_GND transition
can generate IdGnd interrupt from PHY, polling is disabled in RID_A
state and low power mode is allowed.

Change-Id: Ie8b56f70efb3ae8c8d91d2434392962e6723ad51
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>