msm: sdio_cmux: fix corner case in back to back probes
If back to back probes occur because of subsystem restart, and
sdio_open() fails multiple times, sdio_cmux_probe() could attempt to
re-free an already freed workqueue, causing a kernel panic.

The cmux workqueues should not be freed once allocated and the driver
goes into the initialized state.  Change the error handling to correct
that.

Change-Id: Ice16acabb24101195c0169a194c726016e9f69dc
CRs-Fixed: 295466
Signed-off-by: Jeffrey Hugo <jhugo@codeaurora.org>