perf: Fix infinite loop in ARM user perf_event backtrace code
The ARM user backtrace code can get into an infinite loop if it
runs into an invalid stack frame which points back to itself.
This situation has been observed in practice.  Fix it by capping
the number of entries in the backtrace.  This is also what other
architectures do in their backtrace code.

Change-Id: I71509bf81bfe4d4f0f799159adc310e85f0e04ec
Signed-off-by: Sonny Rao <sonnyrao@chromium.org>
Signed-off-by: Jack Cheung <jackc@codeaurora.org>