msm: clock-local: Cleanup function naming conventions
Rename the family of functions with the 'local_clk' prefix to
have an 'rcg_clk' prefix instead, so that the function names
matche the type of clock (and type of struct) those functions
operate on.

This also helps to disambiguate these functions from other
functions that operate on different types of "local clocks"
(ie. branches, PLLs, etc).

Change-Id: I2d5cf6910e5256c8863eb828f5ebba37996f6fd5
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>