msm: audio: qdsp6v2: Support for out-of-band calibration
To support sending of calibration in physical shared memory
the number of voice volume steps is only sent in the first
voice volume calibration block. This means that all products
that do not support out-of-band calibration have to pull
the number of voice volume steps from the first buffer and
place it into the rest of the buffers.

Change-Id: I4d58627f2dc50c8a27b467fb5a7d35da5af7d5a0
Signed-off-by: Ben Romberger <bromberg@codeaurora.org>