defconfig: Enable LOCALVERSION_AUTO and MODVERSIONS
It's nice to have the git hash after the kernel version to easily
identify what commit a crash report comes from. In such a
situation, a matching kernel can be easily built by simply
checking out the hash that is indicated in the kernel version.

Therefore, enable LOCALVERSION_AUTO in all the defconfigs so that
we get the shortened commit hash. Enable MODVERSIONS too,
ensuring that modules won't fail to load when the version of the
kernel changes from one commit to another.

Change-Id: I54b7529a3cecb77459eff1e1c4c5643656f9505d
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>