power: pmic8058-charger: system power control
The pmic charger can control its supply to the system.
The msm_charger uses this facility to enable or disable
the pmic from providing system current.

On a setup with multiple charges, when the external charger
is plugged in it provides the majority of the system current. This reduces
the current drawn from the pmic charger source (usb). If the pmic
were charging the battery, the reduction in the current drawn from
pmic charger source causes a charge completion interrupt in the pmic.
Note that the pmic only monitors the current drawn from usb to signal
end of charging. In this case, the interrupt is fired when the
battery may not be fully charged.
Moreover when the pmic thinks that the battery is fully charged, it
opens up the batfet and prevents any other charger from charging the
battery.
With this fix, when an external charger is plugged in the pmic is asked
to stop providing system current which causes the pmic to close
the batfet.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit 3cee9fb33993dd5bb63081cbe15500f155114cd7)

Change-Id: I9aa981fce6418be15246c86d381225b9265af027
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>