power: pmic8058-charger: chg_done interrupt storm fix
A storm of chg_done interrupt was observed when isl charger
is plugged in.

Register for only the rising edge of that interrupt instead
for both edges.

Also disable the interrupt in the handler - there is no reason
for the interrupt to stay enabled while it waits for a second
for stop_charging to happen.

When the charger is asked to stop charging, it should disable
the interrupts before it cancels the workqueues. There is a
small chance of a race otherwise, where the workqueues are
scheduled after they have been cancelled but before interrupts
are disabled.

Change-Id: I943ad2764a920a32d2a421203cdf91bcf6c2b2ac
Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>