power: pmic8058-charger: chg_done interrupt storm fix
A storm of chg_done interrupt was observed when isl charger
is plugged in.

Register for only the rising edge of that interrupt instead
for both edges.

Also disable the interrupt in the handler - there is no reason
for the interrupt to stay enabled while it waits for a second
for stop_charging to happen.

When the charger is asked to stop charging, it should disable
the interrupts before it cancels the workqueues. There is a
small chance of a race otherwise, where the workqueues are
scheduled after they have been cancelled but before interrupts
are disabled.

Signed-off-by: Abhijeet Dharmapurikar <adharmap@codeaurora.org>
(cherry picked from commit c8b3ae3f4bc2cb964d27c01eb4ef7ec5b95cd591)

Change-Id: I945efe5b6f425c6aefed400b8be6f13809eceb78
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>