CHROMIUM: gobi: protect access to dev->urbs, initialize active
Protect access to dev->urbs, otherwise urbs can be added to the
dev->urbs at the same time that the urbs in a worker thread are
removing urbs.  The could corrupt the list.

Initialize the active urb to NULL, otherwise the module might always
think that an urb is in progress and never submit any urbs thus
blocking all IP traffic.

Signed-off-by: Jason Glasgow <jglasgow@chromium.org>

BUG=chrome-os-partner:4649
TEST=beat on the 3G system.  Connect/Disconnect repeatedly.  Suspend/Resume

Change-Id: Ic560fdc790893c75c3fc3781fbd9126dfcf51128
Reviewed-on: http://gerrit.chromium.org/gerrit/4240
Tested-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>