ASoC: msm8960: changing the ASM closing sequence
Previous ASM closing sequence freed the DMA buffer first and
then stop the ASM session. Since FE is also closed before BE is
closed, it is possible that Q6 ASM continuously gets the audio
data from BE and the data is writing to DMA buffer which doesn't
exist anymore. This causes the Q6 crashes.

CRs-Fixed: 292477
Change-Id: I6c2255e6f2ad2de013945e2209a67ca12bb60d7e
Signed-off-by: Jay Wang <jaywang@codeaurora.org>