usb: msm_otg: Enabling the IDPU in OTGSC register during otg_reset
In some scenarios, as a part of otg_reset the IDPU bit in OTGSC register
is getting disabled. This blocks the ID pin to be sampled and so the
target does not enumerate in the expected mode.

Handling this by setting the IDPU bit in OTGSC register during
otg_reset.

Change-Id: Ibdfd2ed4d5028fe581770803653ac06479bf0833
CRs-Fixed: 296231
Signed-off-by: Rajkumar Raghupathy <raghup@codeaurora.org>