msm: scm: Add atomic SCM APIs
The atomic SCM APIs are useful for commands that are guaranteed
by the secure side to be uninterruptable, atomic and SMP safe.
The calling convention use registers for passing parameters and
return values between the secure and non-secure side. Support
this interface with  scm_call_atomic[1-2]() functions
corresponding to the number of arguments passed.

Change-Id: I151830c4d41fe407bcf38e558bf7a96501f31883
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>