msm: sdio: Change the rx threshold after sleep according to read avail
This fix is required to prevent a latency in ping over
streaming channels in cases where the apps wakes up the modem.
If we restore the old threshold on the wakeup procedure,
the response to the ping is delayed until sdio_al timer
expires.

CRs-fixed: 294068
Signed-off-by: Maya Erez <merez@codeaurora.org>
(cherry picked from commit 2fa8bc482703bebe06905e42251ee1814034a1c3)

Change-Id: I9cef09908a2987dc88b3489a11cbb628196e12eb
Signed-off-by: Raj Kushwaha <rajk@codeaurora.org>