CHROMIUM: ARM: tegra: asymptote: Add new Asymptote init functions
Add new Asymptote init functions and use those for Asymptote boards
as opposed to using the wario/seaboard inits it is currently using.
This required some duplication of the atmel device data since the
I2C address and GPIO numbers are different between the boards.

BUG=None
TEST=Built the kernel and ran it on both seaboard and asymptote. Ensured
the things worked as expected on both boards.

Change-Id: I222c5ca0e3da6a4158b39e22e22cf8a765731575
Signed-off-by: Sean Paul <seanpaul@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/4492
Reviewed-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>