ath6kl: don't force foreground scan when connected
In my setup data transfer stalls when there's data transmission during
scan. After some testing I found out that using background scan
when connected to makes the problem go away. This is more like
a workaround than a proper fix, but as the stall is so severe the
workaround is justified.

With a dual band card this increases scan time when connected from
1.9s to 5.4s. When not connected the scan time is not affected and
is the same 1.9s.

This is a backport of a patch with the same name from ath6kl-cleanup tree.

Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>

BUG=chrome-os-partner:4458
TEST=manual:test normal use on various platforms

Change-Id: Ie51f3f04333c07290c15b949900670c6df81869f
Reviewed-on: http://gerrit.chromium.org/gerrit/5283
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>