CHROMIUM: ath6kl: fix suspend and resume
The current suspend code in ath6kl is not working properly. Most of the
time during response there are sdio errors and sdio stack reloads the
driver.

I was pondering different options how to fix this and decided to take
the simplest route. First remove all the existing suspend/resume code
which was complicated and replace it with code which will leave
the chip powered on but will put it to deep sleep state. Now suspend
works much faster and realiably, the cost being that power consumption
will increase a bit during suspend.

Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>

BUG=chrome-os-partner:4413
TEST=manual:close+open lid with various network configs
TEST=roaming suite

Change-Id: I4ed237c30902499c67fedad9cf04154feb52ebc1
Reviewed-on: http://gerrit.chromium.org/gerrit/4618
Reviewed-by: Sam Leffler <sleffler@chromium.org>
Tested-by: Sam Leffler <sleffler@chromium.org>