CHROMIUM: tegra: harmony: wifi: add board support for WiFi
The SDIO WiFi module requires power from external PMU and 1.2V regulator.
The module gets powered on if - (1) all power sources are enabled, and
(2) power (down) and reset (down) pins are enabled as per spec.

To enable mmc/SDIO driver to detect the WiFi hardware, the WiFi chip has
to be powered-up before mmc driver does probing. So, steps should be as
following: (1) required regulators are on, (2) power/reset of WiFi are
enabled, (3) mmc does probing. Later time, when WiFi driver module is
loaded and registers with SDIO, the SDIO driver knows which H/W the driver
has to be associated with.

BUG=none
TEST=WiFi module is properly powered up and mmc detects it. Sample kernel log:

[    3.202617] mmc0: queuing unknown CIS tuple 0x01 (3 bytes)
[    3.211251] mmc0: queuing unknown CIS tuple 0x1a (5 bytes)
[    3.215018] mmc0: queuing unknown CIS tuple 0x1b (8 bytes)
[    3.215810] mmc0: queuing unknown CIS tuple 0x14 (0 bytes)
[    3.220636] mmc0: queuing unknown CIS tuple 0x80 (1 bytes)
[    3.220894] mmc0: queuing unknown CIS tuple 0x81 (1 bytes)
[    3.221150] mmc0: queuing unknown CIS tuple 0x82 (1 bytes)
[    3.221283] mmc0: new SDIO card at address 0001

Change-Id: I6ff46924913efbeff0a332bd195300a58d2eb8d8
Signed-off-by: Mursalin Akon <makon@nvidia.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/4550
Reviewed-by: Allen Martin <amartin@nvidia.com>
Reviewed-by: Doug Anderson <dianders@chromium.org>