CHROMIUM: video: tegra: Implement EDID query
This change implements the TEGRA_DC_EXT_CONTROL_GET_OUTPUT_EDID ioctl in
the dc_ext interface.

It first adds a way for the tegra dc EDID module to export EDID data
safely, without the risk of reading an incomplete or corrupted EDID in
the presence of hotplug, by moving the actual data to a substructure
with a lifetime maintained by a kref.  Then, that support is plumbed
through the hdmi block (which is currently the only way to get at the
EDID) and out to userspace.

Signed-off-by: Robert Morell <rmorell@nvidia.com>

BUG=chrome-os-partner:4249
TEST=Read the EDID for an HDMI monitor, both through the XRandR
interface and through the debugfs file /sys/kernel/debug/edid1

Change-Id: Icb32a2b08c40340afb9ad6d2d817ac3440265596
Reviewed-on: http://gerrit.chromium.org/gerrit/4632
Tested-by: Robert Morell <rmorell@nvidia.com>
Reviewed-by: Vincent Palatin <vpalatin@chromium.org>
Reviewed-by: Mark Hayter <mdhayter@chromium.org>