Revert "workqueue: add workqueue_empty() api"
This reverts commit bccb069835ef880e437c68a7fed9529c2549605f.

Since the workqueue code deletes the work before executing it,
checking for no work item being currently queued to the workqueue
is not sufficient to guarantee that all the works have finished
execution. Hence, we have moved to using a counter based solution
and so this change is no longer required.

Change-Id: Ib56bd1ce77e8d5e23fffaf3439977d59d931e2a5
Signed-off-by: Pratik Patel <pratikp@codeaurora.org>