usb: gadget: Check for pending requests before priming the ep
For any data to be transferred over usb endpoint, USB S/W
configures the data to be transferred in queue heads and transfer
descriptors and requests h/w to process the data by priming the
endpoint. Sometimes it is observed that endpoint priming is
failing when usb requests are being queued back to back. Although
the h/w is recovering later, some unprocessed requests are
pending in ep requests queue. To avoid this issue, check for
stale requests before priming the endpoint and add the current
request at the endpoint of request queue.

Change-Id: I5fc084b1b7cdf2939372a7f8fb95c9c3995853ca
Signed-off-by: Vamsi Krishna <vskrishn@codeaurora.org>